requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cleo==0.8.1
clikit==0.6.2
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.5
Django==4.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==6.4.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.7.1
h11==0.13.0
html5lib==1.1
idna==3.3
itypes==1.2.0
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.7.0
lockfile==0.12.2
MarkupSafe==2.1.1
msgpack==1.0.4
oauthlib==3.2.0
outcome==1.2.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
django-phonenumber-field[phonenumbers]
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
poetry==1.1.14
poetry-core==1.0.8
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.4.0
pylev==1.4.0
pyparsing==3.0.9
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SecretStorage==3.3.2
selenium==4.4.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomlkit==0.11.1
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.3
webencodings==0.5.1
wsproto==1.2.0
